# docs/openapi.yaml
openapi: 3.1.0
info:
  title: SatL Engine API
  description: |-
    The Docker Engine REST API v1.43 as `satld` implements it -- the only external surface SatL has (invariant #8). Every intentional divergence from Docker carries a numbered entry in `docs/api-compat.md`, cited inline throughout this document.

    The daemon listens on the unix socket `/var/run/satl.sock`. There is no TCP listener for this API: the cluster's own protocol (`satl.internal.v1`) is a separate, node-to-node-only surface on ports 2377 and 2378.

    ## Version prefixes

    Every operation is reachable both on its bare path (`/info`) and under a Docker version prefix (`/v1.43/info`). Prefixes are accepted from 1.24 up to 1.43; a prefix outside that range is a `400` in Docker's error shape, not a 404. A leading segment that merely starts with `v` but is not `v<major>.<minor>` -- `/v1/info` -- is not version negotiation at all and routes (and 404s) as an ordinary path, where Docker would parse `1` as a version.

    ## Responses this document cannot attach to an operation

    - **Unknown path**: `404` with the body `{"message": "page not found"}`. Every Docker endpoint SatL does not implement answers this shape -- attach, commit, export, rename, restart, pause, update, top, stats, changes, archive, build, push, and the rest of the `/images/{name}/*` family (api-compat #22, #144, #152).
...
    - **Out-of-range version prefix**: `400` with Docker's error body, minted before routing.

...

...

    ## Error shape
...
servers:
...
paths:
...
  /containers/{id}/exec:
    post:
      tags:
      - Exec
      summary: '`POST /containers/{id}/exec`.'
      description: 'Creates an exec instance. Exec is non-interactive: `Tty:true` is a 400 on create *and* on start, `Privileged:true` is a 400, `DetachKeys` is ignored, and stdin on the hijacked socket is read and discarded (api-compat #17, #38).'
      operationId: ContainerExec
      parameters:
      - name: id
        in: path
        description: Container (task) ID or name.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecCreateBody'
        required: true
      responses:
        '201':
          description: The exec instance was created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecCreateResponse'
        '400':
          description: '`Tty` or `Privileged` was set, or the body is invalid.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: No such container.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: The container is not running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: Not implemented by this daemon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /events:
    get:
      tags:
      - System
      summary: '`GET /events?since=&until=&filters=`.'
      description: '**Newline-delimited JSON, not a JSON array.** The 200 body is a chunked stream carrying one JSON object per line, terminated by `\n`, written as each event happens; a client must read it line by line and never buffer it as a single document. The schema below describes one such line. `filters` is not read at all -- no key is validated either, so a nonsense filter passes as silently as a real one (api-compat #21); `?since=` is accepted but there is no replayable history (api-compat #37), and `?until=` is a 501. The legacy top-level `status`/`id`/`from` fields are omitted.'
      operationId: SystemEvents
      parameters:
      - name: since
        in: query
        description: 'Accepted and parsed for validity; there is no replayable history (api-compat #37).'
        required: false
        schema:
          type: string
      - name: until
        in: query
        description: 'Rejected with 501 (api-compat #21).'
        required: false
        schema:
          type: string
      - name: filters
        in: query
        description: 'Not read at all (api-compat #21).'
        required: false
        schema:
          type: string
      responses:
        '200':
          description: One JSON object per line, not a JSON array.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          description: '`?since=` is not a timestamp.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: '`?until=` was set, or the daemon has no event source wired.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /exec/{id}/json:
    get:
      tags:
      - Exec
      summary: '`GET /exec/{id}/json`.'
      description: '`ProcessConfig.privileged` is always false, `DetachKeys` always empty and `CanRemove` always true (api-compat #19).'
      operationId: ExecInspect
      parameters:
      - name: id
        in: path
        description: Exec instance ID.
        required: true
        schema:
          type: string
      responses:
        '200':
          description: The exec instance document.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecInspectResponse'
        '404':
          description: No such exec instance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: Not implemented by this daemon.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /exec/{id}/start:
    post:
      tags:
      - Exec
      summary: '`POST /exec/{id}/start`.'
      description: '**Not JSON, and two shapes.** With `Connection: Upgrade` and `Upgrade: tcp` the daemon answers `101 Switching Protocols` with an empty body and then owns the raw socket, writing Docker multiplexed frames over it until the process exits. **Without** the upgrade headers it answers `200` with the same frames as an ordinary chunked body, where Docker would hijack in both cases (api-compat #18). Output is delivered when the process exits rather than streamed live (api-compat #38).'
      operationId: ExecStart
      parameters:
      - name: id
        in: path
        description: Exec instance ID.
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecStartBody'
        required: true
      responses:
        '101':
          description: 'Connection hijacked: the raw socket now carries Docker multiplexed frames.'
        '200':
          description: 'No upgrade was requested: the same frames as a chunked body (api-compat #18).'
          content:
            application/vnd.docker.raw-stream:
              schema:
                type: string
        '400':
          description: '`Tty` was set, or the body is invalid.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: No such exec instance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: The exec instance has already run.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Daemon error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '501':
          description: '`Detach` was set (api-compat #17), or the daemon has no executor wired.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '503':
          description: This node is not a swarm manager, or no manager is reachable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
components:
  schemas:
...
    ErrorBody:
      type: object
      description: 'Docker error envelope: every non-2xx response body is `{"message": "..."}`.'
      required:
      - message
      properties:
        message:
          type: string
          description: Human-readable error message.
    EventActorResponse:
      type: object
      description: '`Actor` of an [`EventResponse`].'
      required:
      - ID
      - Attributes
      properties:
        Attributes:
          type: object
          description: Free-form attributes.
          additionalProperties:
            type: string
          propertyNames:
            type: string
        ID:
          type: string
          description: Object ID.
    EventResponse:
      type: object
      description: One `GET /events` message.
      required:
      - Type
      - Action
      - Actor
      - scope
      - time
      - timeNano
      properties:
        Action:
          type: string
          description: What happened.
        Actor:
          $ref: '#/components/schemas/EventActorResponse'
          description: Who it happened to.
        Type:
          type: string
          description: Object kind (`container`, `image`, `volume`, `network`).
        scope:
          type: string
          description: '`local` or `swarm`.'
        time:
          type: integer
          format: int64
          description: Event time, unix seconds.
        timeNano:
          type: integer
          format: int64
          description: Event time, unix nanoseconds.
    ExecCreateBody:
      type: object
      description: '`POST /containers/{id}/exec` body.'
      properties:
        AttachStderr:
          type: boolean
          description: Stream stderr back.
          default: false
        AttachStdin:
          type: boolean
          description: Attach stdin (accepted, then discarded).
          default: false
        AttachStdout:
          type: boolean
          description: Stream stdout back.
          default: false
        Cmd:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/StringOrList'
            description: Command and arguments.
          default: null
        DetachKeys:
          type:
          - string
          - 'null'
          description: Detach key sequence (ignored — SatL never attaches a TTY).
          default: null
        Env:
          type:
          - array
          - 'null'
          items:
            type: string
          description: Extra environment.
          default: null
        Privileged:
          type: boolean
          description: Run privileged (rejected).
          default: false
        Tty:
          type: boolean
          description: Allocate a TTY (rejected in M1).
          default: false
        User:
          type:
          - string
          - 'null'
          description: User to run as.
          default: null
        WorkingDir:
          type:
          - string
          - 'null'
          description: Working directory.
          default: null
    ExecCreateResponse:
      type: object
      description: '`POST /containers/{id}/exec` response.'
      required:
      - Id
      properties:
        Id:
          type: string
          description: Exec instance ID.
    ExecInspectResponse:
      type: object
      description: '`GET /exec/{id}/json`.'
      required:
      - ID
      - ContainerID
      - Running
      - ExitCode
      - Pid
      - OpenStdin
      - OpenStdout
      - OpenStderr
      - CanRemove
      - DetachKeys
      - ProcessConfig
      properties:
        CanRemove:
          type: boolean
          description: Whether the instance can be removed (always true).
        ContainerID:
          type: string
          description: Container the exec belongs to.
        DetachKeys:
          type: string
          description: Detach key sequence — always empty.
        ExitCode:
          type: integer
          format: int64
          description: Exit code, 0 while running.
        ID:
          type: string
          description: Exec instance ID.
        OpenStderr:
          type: boolean
          description: Whether stderr was attached.
        OpenStdin:
          type: boolean
          description: Whether stdin was attached.
        OpenStdout:
          type: boolean
          description: Whether stdout was attached.
        Pid:
          type: integer
          format: int64
          description: PID inside the jail, 0 when unknown.
        ProcessConfig:
          $ref: '#/components/schemas/ProcessConfig'
          description: The process being run.
        Running:
          type: boolean
          description: Whether the process is still running.
    ExecStartBody:
      type: object
      description: '`POST /exec/{id}/start` body.'
      properties:
        Detach:
          type: boolean
          description: Run without streaming output back (rejected in M1).
          default: false
        Tty:
          type: boolean
          description: Allocate a TTY (rejected in M1).
          default: false
...
    ProcessConfig:
      type: object
      description: |-
        `ProcessConfig` of an exec inspect document (lower-case keys, as Docker
        serializes them).
      required:
      - tty
      - entrypoint
      - arguments
      - privileged
      - user
      properties:
        arguments:
          type: array
          items:
            type: string
          description: Its arguments.
        entrypoint:
          type: string
          description: The command.
        privileged:
          type: boolean
          description: Whether the exec is privileged (always false).
        tty:
          type: boolean
          description: Whether a TTY was requested.
        user:
          type: string
          description: User the process runs as.
...
    StringOrList:
      oneOf:
      - type: string
        description: A bare string — a shell-style single argument.
      - type: array
        items:
          type: string
        description: The usual exec-form list.
      description: |-
        A Docker `StrSlice`: JSON accepts either a single string or an array of
        strings for `Cmd` and `Entrypoint`.
...